% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{supp_annotation}
\alias{supp_annotation}
\title{Expand annotation using phyloseq tables}
\usage{
supp_annotation(mvar_object, physeq, n1 = "species", n2 = "site")
}
\description{
Expand annotation using phyloseq tables
}

